POC to improve test readability with project test file
